requirements.txt
streamlit>=1.24.0
langchain>=0.0.252
boto3
botocore
python-dotenv
sqlalchemy>=1.4.47
snowflake-sqlalchemy
sqlalchemy-aurora-data-api
PyAthena[SQLAlchemy]==2.25.2
anthropic
redshift-connector==2.0.910
sqlalchemy-redshift==0.8.14
pexpect
langchain_experimental
gql
httpx
requests-toolbelt
snowflake-sqlalchemy
flask-sqlalchemy 
openai